TST: moving fixtures to conftest.py
